fix: invalidatae table positions if repaired
