+++ __tests__/parsers/tables.test.ts
@@ -518,300 +518,12 @@ const x = 1;
       expect(cells).toHaveLength(2);
     });
 
-    describe('unclosed JSX inside cells', () => {
-      it('repairs an unclosed inline element inside a cell', () => {
-        const doc = `<Table>
-  <thead><tr><th>Heading</th></tr></thead>
-  <tbody>
-    <tr>
-      <td>cell with <span>unclosed inline</td>
-    </tr>
-  </tbody>
-</Table>`;
-
-        const { tree } = parseMdxishWithSource(doc);
-        const tableNode = collectNodes(tree, 'table');
-        expect(tableNode).toHaveLength(1);
-
-        const hast = mdxish(doc);
-        const tables = findAllElementsByTagName(hast, 'table');
-        expect(tables).toHaveLength(1);
-
-        const cells = findAllElementsByTagName(tables[0], 'td');
-        expect(cells).toHaveLength(1);
-        expect(JSON.stringify(cells[0])).toContain('unclosed inline');
-      });
-
-      it('repairs an unclosed td before another td', () => {
-        const doc = `<Table>
-  <thead><tr><th>A</th><th>B</th></tr></thead>
-  <tbody>
-    <tr>
-      <td>first
-      <td>second</td>
-    </tr>
-  </tbody>
-</Table>`;
-
-        const { tree } = parseMdxishWithSource(doc);
-        const tableNode = collectNodes(tree, 'table');
-        expect(tableNode).toHaveLength(1);
-
-        const hast = mdxish(doc);
-        const tables = findAllElementsByTagName(hast, 'table');
-        expect(tables).toHaveLength(1);
-
-        const cells = findAllElementsByTagName(tables[0], 'td');
-        expect(cells.length).toBeGreaterThanOrEqual(2);
-      });
-
-      it('repairs unclosed jsx alongside a jsx expression attribute', () => {
-        const doc = `<Table>
-  <thead><tr><th style={{ textAlign: "left" }}>Heading</th></tr></thead>
-  <tbody>
-    <tr>
-      <td>this is <em>broken emphasis</td>
-    </tr>
-  </tbody>
-</Table>`;
-
-        const hast = mdxish(doc);
-        const tables = findAllElementsByTagName(hast, 'table');
-        expect(tables).toHaveLength(1);
-
-        const cells = findAllElementsByTagName(tables[0], 'td');
-        expect(cells).toHaveLength(1);
-        expect(JSON.stringify(cells[0])).toContain('broken emphasis');
-      });
-
-      it('still parses markdown inside a cell after repair', () => {
-        // The unclosed <em> would normally crash remarkMdx. After repair the
-        // table goes through the MDX-aware pipeline, which means **bold**
-        // inside the (well-formed) sibling cell still becomes a <strong>.
-        const doc = `<Table>
-  <thead><tr><th>A</th><th>B</th></tr></thead>
-  <tbody>
-    <tr>
-      <td>before <em>oops</td>
-      <td>**emphasized**</td>
-    </tr>
-  </tbody>
-</Table>`;
-
-        const hast = mdxish(doc);
-        const strongs = findAllElementsByTagName(hast, 'strong');
-        expect(strongs.length).toBeGreaterThan(0);
-        expect(JSON.stringify(strongs[0])).toContain('emphasized');
-      });
-
-      it('inserts the closer at a blank line so it lands in the same paragraph as the open', () => {
-        // <span> opens in the first paragraph but never closes; the next paragraph
-        // continues with more text before </td>. MDX requires the synthetic </span>
-        // to land at the paragraph boundary (the blank line), not at </td>, or it
-        // gets parsed as a different paragraph and remarkMdx still throws.
-        const doc = `<Table>
-  <thead><tr><th>Description</th></tr></thead>
-  <tbody>
-    <tr>
-      <td>
-        <span style="color:red">first paragraph with unclosed span.
-
-        second paragraph keeps going.
-      </td>
-    </tr>
-  </tbody>
-</Table>`;
-
-        const { tree } = parseMdxishWithSource(doc);
-        const tableNode = collectNodes(tree, 'table');
-        expect(tableNode).toHaveLength(1);
-
-        const hast = mdxish(doc);
-        const tables = findAllElementsByTagName(hast, 'table');
-        expect(tables).toHaveLength(1);
-
-        const cells = findAllElementsByTagName(tables[0], 'td');
-        expect(cells).toHaveLength(1);
-        const json = JSON.stringify(cells[0]);
-        expect(json).toContain('first paragraph');
-        expect(json).toContain('second paragraph');
-      });
-
-      it('table does not break with <br> tag', () => {
-        const doc = `<Table>
-  <thead><tr><th>A</th><th>B</th></tr></thead>
-  <tbody>
-    <tr>
-      <td>before <br> oops</td>
-    </tr>
-  </tbody>
-</Table>`;
-
-        const { tree } = parseMdxishWithSource(doc);
-        const tableNode = collectNodes(tree, 'table');
-        expect(tableNode).toHaveLength(1);
-      });
-
-      it('table does not break when there is <object> tag', () => {
-        const doc = `<Table>
-    <thead>
-      <tr>
-        <th>
-          Parameter
-        </th>
-
-        <th>
-          Notes
-        </th>
-      </tr>
-    </thead>
-
-    <tbody>
-      <tr>
-        <td>
-          Array <object>
-        </td>
-
-        <td>
-          Each object can have:<ul><li>type (Enum:
-   ROLLBACK)</li><li>enabled (default: true)</li></ul><br
-  /><br />
-        </td>
-      </tr>
-    </tbody>
-  </Table>`;
-
-        const { tree } = parseMdxishWithSource(doc);
-        const tableNode = collectNodes(tree, 'table');
-        expect(tableNode).toHaveLength(1);
-      });
-
-      it('table does not break when there is genuine escaped HTML tag', () => {
-        const doc = `<Table>
-  <thead>
-    <tr>
-      <th>
-        Parameter
-      </th>
-
-      <th>
-        Notes
-      </th>
-
-      <th>
-        Type
-      </th>
-
-      <th>
-        Required
-      </th>
-
-      <th>
-        Possible Values
-      </th>
-    </tr>
-  </thead>
-
-  <tbody>
-    <tr>
-      <td>
-        urlTracker
-      </td>
-
-      <td>
-        Specifies
-
-        **_Note:_**_&#x20;Somehthing_<br />**The urlTracker objects are detailed** <p><a href="#C4">**here**</a></p>
-      </td>
-
-      <td>
-        string
-      </td>
-
-      <td>
-
-      </td>
-
-      <td>
-
-      </td>
-    </tr>
-
-    <tr>
-      <td>
-        badgeSettings
-      </td>
-
-      <td>
-        Cell
-      </td>
-
-      <td>
-        Array <object>
-      </td>
-
-      <td>
-        N
-      </td>
-
-      <td>
-        Enum Values: ROLLBACK, EXPRESS\\_DELIVERY
-      </td>
-    </tr>
-
-    <tr>
-      <td>
-        associatedItems
-      </td>
-
-      <td>
-        Something
-      </td>
-
-      <td>
-        Array \\<string>
-      </td>
-
-      <td>
-        N
-      </td>
-
-      <td>
-        valid <<variable>>
-      </td>
-    </tr>
-  </tbody>
-</Table>`;
-
-        const { tree } = parseMdxishWithSource(doc);
-        const tableNode = collectNodes(tree, 'table');
-        expect(tableNode).toHaveLength(1);
-      });
-
-      it('does not modify well-formed tables (repair is a no-op)', () => {
-        const doc = `<Table>
-  <thead><tr><th>Heading</th></tr></thead>
-  <tbody>
-    <tr><td>fine</td></tr>
-  </tbody>
-</Table>`;
-
-        const hast = mdxish(doc);
-        const tables = findAllElementsByTagName(hast, 'table');
-        expect(tables).toHaveLength(1);
-
-        const cells = findAllElementsByTagName(tables[0], 'td');
-        expect(cells).toHaveLength(1);
-      });
-    });
-
     describe('asymmetric inline/flow JSX inside cells', () => {
       // mdxjs throws when a JSX element's opener has trailing text on its line
       // but its closer sits alone on its own line (or vice versa). Without
       // recovery the html block falls through unparsed and *everything* inside
       // the table — including unrelated cells with markdown links — fails to
       // render. These tests pin the normalizer.
-
       it('handles opener with trailing content and closer alone', () => {
         const doc = `<Table>
   <thead><tr><th>H1</th><th>H2</th></tr></thead>
@@ -1396,11 +1108,6 @@ None of the following content will get rendered!`;
   });
 
   describe('jsx tables with legacy variables', () => {
-    // Without masking `<<NAME>>` in the repair/normalize paths, htmlparser2
-    // would see the inner `<NAME>` as a tag opener with no close and either
-    // inject a phantom `</NAME>` (repair) or insert disruptive newlines
-    // (normalize), corrupting the table when *another* cell triggers the
-    // retry path. Mask the variable so it never reaches the tag tokenizer.
     it('preserves <<variable>> when another cell triggers the malformed-retry path', () => {
       const doc = `<Table>
   <thead>
@@ -1427,8 +1134,6 @@ None of the following content will get rendered!`;
       expect(variables[0].properties).toMatchObject({ name: 'NAME', isLegacy: true });
     });
 
-    // Malformed `<<string>` (single trailing `>`) should not be parsed as a
-    // phantom <string> tag opener even when the malformed-retry path runs.
     it('does not treat <<string> as a phantom tag during malformed-retry', () => {
       const doc = `<Table>
   <thead>
